Adding a new automation program
This code will grab events from the EDR corresponding with endpoint activity data. User input is the hostname of the host you want to query. It will then query that hostname for any endpoint activity data that contains a "logonUser" field in the JSON response. This is a method of tracking user activity on an endpoint from the EDR because processes that the EDR tracks, like "svchost.exe" are associated with the current signed in user. This should let you see what users have been using the endpoint in the timeframe specified. 

I may need to add a parameter that specifies the number of results returned because the default is to return 500, and the base query in the EDR itself returns thousands of results in just a 24 hour span.