Update Papercut.py
I've edited the code so that now it will make several API calls, each calling for the last 5000 events in a half hour increment counting backwards for 24 hours from runtime. Then it'll grab unique logonUser fields and store them in the output.txt file

This should be a relatively effective way of collecting what user was on a machine around what time.